Clean up the API client module:
- Break up the file where all client methods have been defined (it has grown too big). Separate the methods into 4 files concerned with projects, repositories, kus and users respectively. However, I kept the method names the same for the time being (`client.getProjects()` is still valid, no `client.projects.list()` or so...) to reduce the scope of the refactoring.
- Rename from `gitlab` to `api-client` as we are not querying GitLab directly anymore.
- Do not alert errors per default anymore. At this stage errors should be handled by the application more gracefully.
- Parametrize responseType of renkuFetch
- Abandon python-inspired underscore-syntax as it has been used inconsistently.
