+++ src/merge-request/MergeRequest.container.js
@@ -19,7 +19,7 @@
 import React, { Component } from 'react';
 import { NotebookComparisonPresent, MergeRequestPresent } from './MergeRequest.present';
 import Notebook from '../file/Notebook'
-import { ACCESS_LEVELS } from '../gitlab';
+import { ACCESS_LEVELS } from '../api-client';
 
 class MergeRequestContainer extends Component {
   constructor(props){
